pyproject.toml
[project]
name = "hammond"
version = "0.1.0"
description = "Misc automations for my life"
requires-python = ">=3.13"
dependencies = [
    "coloredlogs>=15.0.1",
    "discord-py>=2.6.4",
    "hypercorn>=0.17.3",
    "pydantic>=2.12.4",
    "requests>=2.32.5",
    "restate-sdk[serde]>=0.11.0",
    "urlextract>=1.9.0",
]

[project.scripts]
hammond = "hammond.main:main"

[build-system]
requires = ["hatchling", "editables"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "types-requests>=2.32.4.20250913",
]

[tool.pyright]
executionEnvironments = [
  { root = "tests", reportUnusedFunction = false, reportUnusedVariable = false, reportUnknownParameterType = false, reportMissingParameterType = false, reportUnusedParameter = false }
]
stubPath = "stubs"

[tool.mypy]
namespace_packages = true
explicit_package_bases = true
mypy_path = "src:stubs"

[[tool.mypy.overrides]]
module = "coloredlogs"
ignore_missing_imports = true
